update notifications websocket

+++ src/routes/v1/User.ts
@@ -1,10 +1,11 @@
 import { Elysia, t } from 'elysia';
-import { PrismaClient, usage_status } from '@prisma/client';
+import { PrismaClient, users_usage_status } from '@prisma/client';
 import { jwt } from '@elysiajs/jwt';
 import { unlink } from "node:fs/promises";
 import path from 'path';
 import { getThaiDate, JWTPayloadUser } from '../../../lib/lib';
-
+import { existsSync } from 'node:fs';
+import cors from '@elysiajs/cors';
 
 const prisma = new PrismaClient();
 const SECRET_KEY = process.env.SECRET_KEY;
@@ -14,19 +15,8 @@ if (!SECRET_KEY) {
 }
 
 const app = new Elysia()
+    .use(cors())
     .use(jwt({ name: 'jwt', secret: SECRET_KEY }))
-    .get("/", async ({ set }) => {
-        try {
-            const users = await prisma.users.findMany()
-            return users;
-        } catch (error) {
-            set.status = 500;
-            return {
-                success: false,
-                message: `Something went wrong: ${(error as Error).message}`
-            };
-        }
-    })
     .derive(async ({ headers, jwt, set }) => {
         const authHeader = headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
@@ -40,6 +30,31 @@ const app = new Elysia()
             where: { id: payloadUser.id },
         });
         return { payloadUser, existingUser }
+    })
+    .post('/admin/notify', ({ body, set }) => {
+        app.server!.publish('notifications', JSON.stringify({
+            type: 'notification',
+            content: body.body
+        }))
+
+        return { success: true, message: 'Notification sent' }
+    }, {
+        body: t.Object({
+            title: t.String(),
+            body: t.String()
+        })
+    })
+    .get("/", async ({ set }) => {
+        try {
+            const users = await prisma.users.findMany()
+            return users;
+        } catch (error) {
+            set.status = 500;
+            return {
+                success: false,
+                message: `Something went wrong: ${(error as Error).message}`
+            };
+        }
     })
     .get('/me', async ({ set, payloadUser }) => {
         if (!payloadUser) {
@@ -90,25 +105,16 @@ const app = new Elysia()
         }
 
         if (profile_picture) {
-            const file = profile_picture;
             if (existingUser.profile_picture !== "default-profile.jpg") {
                 const oldFilePath = `public/images/user_images/${existingUser.profile_picture}`;
-                try {
+                if (existsSync(oldFilePath)) {
                     await unlink(oldFilePath);
-                } catch (error) {
-                    console.error(`เกิดข้อผิดพลาดในการลบรูปโปรไฟล์เก่า: ${error}`);
                 }
             }
-
-            try {
-                const fileName = `${payloadUser.id}-${Date.now()}${path.extname(file.name)}`;
-                const filePath = `public/images/user_images/${fileName}`;
-                await Bun.write(filePath, await file.arrayBuffer());
-                profile_pictureName = fileName;
-            } catch (error) {
-                console.error(`เกิดข้อผิดพลาดในการอัพโหลดรูปโปรไฟล์ใหม่: ${error}`);
-                return ({ success: false, message: "ไม่สามารถอัพโหลดรูปโปรไฟล์ใหม่ได้" });
-            }
+            const fileName = `${payloadUser.id}-${Date.now()}${path.extname(profile_picture.name)}`;
+            const filePath = `public/images/user_images/${fileName}`;
+            await Bun.write(filePath, await profile_picture.arrayBuffer());
+            profile_pictureName = fileName;
         }
 
         try {
@@ -118,15 +124,30 @@ const app = new Elysia()
                 updateData.profile_picture = profile_pictureName;
             }
 
-            if (currentPassword && newPassword) {
-                const isPasswordValid = await Bun.password.verify(currentPassword, existingUser.password);
-                if (!isPasswordValid) { set.status = 400; return { success: false, message: "รหัสผ่านปัจจุบันไม่ถูกต้อง" }; }
+            try {
+                if (currentPassword && newPassword) {
+                    if (currentPassword && !newPassword) {
+                        throw new Error("กรุณากรอกรหัสผ่านใหม่ที่ต้องการเปลี่ยน")
+                    } else if (!currentPassword && newPassword) {
+                        throw new Error("กรุณากรอกรหัสผ่านปัจจุบันของคุณ")
+                    }
+                    const isPasswordValid = await Bun.password.verify(currentPassword, existingUser.password);
+                    if (!isPasswordValid) {
+                        throw new Error("รหัสผ่านปัจจุบันไม่ถูกต้อง")
+                    }
 
-                const hashedPassword = await Bun.password.hash(newPassword, {
-                    algorithm: "bcrypt",
-                    cost: 4,
-                });
-                updateData.password = hashedPassword;
+                    const hashedPassword = await Bun.password.hash(newPassword, {
+                        algorithm: "bcrypt",
+                        cost: 4,
+                    });
+                    updateData.password = hashedPassword;
+                }
+            } catch (error) {
+                set.status = 400;
+                return {
+                    success: false,
+                    message: error
+                };
             }
 
             const updatedUser = await prisma.users.update({
@@ -142,7 +163,6 @@ const app = new Elysia()
                 user: userWithoutPassword
             };
         } catch (error) {
-            console.log(error);
             set.status = 500;
             return {
                 success: false,
@@ -151,16 +171,16 @@ const app = new Elysia()
         }
     }, {
         body: t.Object({
-            firstname: t.String(),
-            lastname: t.String(),
-            email: t.String(),
-            tel: t.String(),
+            firstname: t.Optional(t.String()),
+            lastname: t.Optional(t.String()),
+            email: t.Optional(t.String()),
+            tel: t.Optional(t.String()),
             profile_picture: t.Optional(t.File()),
             currentPassword: t.Optional(t.String()),
             newPassword: t.Optional(t.String())
         })
     })
-    .put("/update-user-status/:status", async ({ headers, jwt, set, params: { status }, payloadUser }) => {
+    .put("/update-user-status/:status", async ({ set, params: { status }, payloadUser }) => {
         if (!payloadUser) {
             set.status = 401;
             return { success: false, message: "token ไม่ถูกต้อง" };
@@ -182,6 +202,7 @@ const app = new Elysia()
                     message: "ตัวตรวจสอบล้มเหลว"
                 };
             }
+
             set.status = 200;
             return {
                 success: true,
@@ -196,9 +217,8 @@ const app = new Elysia()
         }
     }, {
         params: t.Object({
-            status: t.Enum(usage_status),
+            status: t.Enum(users_usage_status),
         })
     })
-
 
 export default app;
